feat(ui): extract sub-components, add shadcn UI, and set up DS tokens
- Add shadcn/ui primitives: badge, button, card, select, sheet, skeleton
- Extract dashboard: BalanceCard, CategoryBreakdownCard, NoCoupleState, NewInstancesBanner
- Extract history: MonthCard, MonthDetail
- Extract settings: NoCoupleCard, PendingInvitationsCard
- Refactor pages to consume new components
- Replace shared Badge with shadcn ui/badge; update FAB and install-prompt
- Add DS token variables to globals.css (accent-foreground, status-*, border-*)
- Fix mojibake in 4 pages (Spanish accented chars corrupted from encoding issue)
- Repair handleCreateCouple/handleInvite corruption in settings/page.tsx
- Replace color:white with var(--accent-foreground) on accent buttons
