+++ package.json
@@ -31,6 +31,7 @@
     "lucide-react": "^1.11.0",
     "next": "16.2.4",
     "next-themes": "^0.4.6",
+    "radix-ui": "^1.4.3",
     "react": "19.2.4",
     "react-dom": "19.2.4",
     "react-hook-form": "^7.74.0",
@@ -59,6 +60,7 @@
     "lint-staged": "^16.4.0",
     "prettier": "^3.8.3",
     "prettier-plugin-tailwindcss": "^0.7.3",
+    "shadcn": "^4.6.0",
     "supabase": "^2.95.3",
     "tailwindcss": "^4",
     "typescript": "^6.0.3",
